refactor!: The view model the FragmentEventDetails.kt is now delegated to activityViewModels to share variable with other fragment such as CalendarFragment.kt.
refactor!: The CalendarOptionsFragment.kt has been created and replace the old CalendarSettingsFragment.kt which is now PreferencesFragment.kt.

feat: The Database has now a new table which is courses which stores the courses visibility (which can be edited in the CalendarOptionsFragment.kt by the user via the UI).

ui: Multiple icon have been added to the project such as "check" icons which are used in the CalendarOptionsFragment.kt

ui: A floating action button has been added to allow the user to configure the application. It redirects to the PreferencesFragment.kt.

ui: The hidden layout in the CalendarFragment which appears on the right is now dedicated to the Calendar options such as hiding courses or setting the calendar link ( which will be done in a next update ).

style: The android:colorBackground setting has been edited to match the dark theme.