refactor: Function get inside PreferencesManager.kt has now a default parameter which allows to automatically infer the return type.
refactor: The serialize and deserialize throw a PreferenceCastException when the preference hasn't been successfully casted, this permit to detect preferences errors or incompatibilities with older versions.

fix: current event is set to null when the event details are folded to avoid the card to re-pop when the Fragment is restarted.

fix: EventDetails loading isn't buggy anymore.
fix: Note isn't deleted when the reminder isn't set.
fix: Notes are deleted when empty.