refactor: nullable types have been removed from SimplePreference.kt. As they can be done with the classic Converter and Delegate classes present in SimplePreference they have been removed. Also the implementations in SimplePreference.kt have been corrected.
refactor: Some code refactored in PreferencesManager.kt.
Some classes and function have been lifted out of the file to clean up the code and make it more readable.

feat: The School class is now annotated with @Serializable.
As the converter used to Serialize/Deserialize the School class has changed to use Kotlin's serialization library, it was necessary to do this update.