improvement: PreferenceKeys holds now an Integer pointing to a string resource.
fix: All usages of PreferenceKeys have been updated to match the new implementation.

Previous implementation was using String to store the Preference keys. This was leading to errors while changing them because almost all the time, changing them into the preferences.xml file was forgot by the programmer. A getter has been added to get the resource with a Context.