refactor: background_services.Updater renamed to UpdateScheduler.kt refactor: Update delegated to Updater.kt.
To let the application evolve in a way that we can add the updater we want without increasing complexity and classes dependencies, Updater is now UpdateScheduler.kt and will only schedule updates. On the other hand, the new interface Updater.kt is designed to create new Updater on demand. UpdaterFactory.kt will now handle the creation of updater given the value passed to it.