upgrade authentication system (#32)
fix(connection): update to match the new authentication system

ref!: upgrade to kotlin version 1.8.22

ref!: switching to viewBinding

This commit upgrade the authentication behavior to match the new university's authentication system.  
As old version of KTor was bugged, the version has been upgraded and so, the Kotlin version too.  
Those upgrades creates incompatibilities with the old view system and so, it has been migrated to the new viewbinding model.