The role dropdown now "remembers" the current setting
The problem was, handleCredentialsUpdate() would trigger a credential update on the ViewModel through the onChange listeners, which in turn would mess up all the other fields. The problem didn't arise before because the TextFields would not accept an empty field a new value, but the dropdown has a default value instead.

With this fix, there is now an utility method to prevent reactivity when modifying the value programmatically. I don't think it can introduce race conditions since the observer events happen on the UI thread as well.